requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
Flask==3.0.3
gotrue==2.5.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
postgrest==0.16.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.3
supafunc==0.4.7
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.3
